Ate the bugs.
In app/main.c, there were two issues. udi_cdc_getc was being called after host_rx init, which won't work, because host_rx intercepts all incoming characters. The second bug was not setting up the physical layer for the sha204p chip. The fix was to call board_init from sha204/board.c. However, there was a name clash, so that function was renamed to sha204_board_init.

In sha204/main.c, there was no infinite loop for the code to fall into at the end of execution, leading to the weird partial print issues and so on.